Fix missing ticket medium on second scan attempt
A ticket might be invalid on the first scan attempt, because it is still too early.
On a later second attempt it might be valid, so it gets returned from the ticketsByBarcode cache,
but now the signed info is no longer available. The info about the ticket medium is therefore lost
and results in the bug described in a307e2d. This commits fixes this.